Move LoginWithDeviceComponent to libs
